Rename NumPy demo modules
