+++ tests/transmogrifier/test_ascii_raster_bounds.py
@@ -1,7 +1,7 @@
 import numpy as np
 from types import SimpleNamespace
 
-from src.cells.softbody.demo.run_numpy_demo import _rasterize_ascii_numpy
+from src.cells.softbody.demo.numpy_sim_coordinator import _rasterize_ascii_numpy
 from src.cells.cellsim.mechanics.softbody0d import Softbody0DProvider, SoftbodyProviderCfg
 from src.cells.cellsim.data.state import Cell, Bath
